Split `protocol.pyi` into submodules
