+++ parsec/_parsec_pyi/protocol/cmd.pyi
@@ -1,0 +1,17 @@
+# Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
+from typing import Any
+
+class AuthenticatedAnyCmdReq:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> Any: ...
+
+class InvitedAnyCmdReq:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> Any: ...
+
+class AnonymousAnyCmdReq:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> Any: ...
