+++ parsec/_parsec_pyi/protocol/organization.pyi
@@ -1,0 +1,84 @@
+# Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
+from typing import Iterable
+
+from parsec._parsec import UsersPerProfileDetailItem
+
+# Organization
+class OrganizationStatsReq:
+    def dump(self) -> bytes: ...
+
+class OrganizationStatsRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> OrganizationStatsRep: ...
+
+class OrganizationStatsRepOk(OrganizationStatsRep):
+    def __init__(
+        self,
+        data_size: int,
+        metadata_size: int,
+        realms: int,
+        users: int,
+        active_users: int,
+        users_per_profile_detail: Iterable[UsersPerProfileDetailItem],
+    ) -> None: ...
+    @property
+    def data_size(self) -> int: ...
+    @property
+    def metadata_size(self) -> int: ...
+    @property
+    def realms(self) -> int: ...
+    @property
+    def users(self) -> int: ...
+    @property
+    def active_users(self) -> int: ...
+    @property
+    def users_per_profile_detail(self) -> tuple[UsersPerProfileDetailItem]: ...
+
+class OrganizationStatsRepNotAllowed(OrganizationStatsRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class OrganizationStatsRepNotFound(OrganizationStatsRep): ...
+
+class OrganizationStatsRepUnknownStatus(OrganizationStatsRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class OrganizationConfigReq:
+    def dump(self) -> bytes: ...
+
+class OrganizationConfigRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> OrganizationConfigRep: ...
+
+class OrganizationConfigRepOk(OrganizationConfigRep):
+    def __init__(
+        self,
+        user_profile_outsider_allowed: bool,
+        active_users_limit: int | None,
+        sequester_authority_certificate: bytes | None,
+        sequester_services_certificates: Iterable[bytes] | None,
+    ) -> None: ...
+    @property
+    def user_profile_outsider_allowed(self) -> bool: ...
+    @property
+    def active_users_limit(self) -> int | None: ...
+    @property
+    def sequester_authority_certificate(self) -> bytes | None: ...
+    @property
+    def sequester_services_certificates(self) -> tuple[bytes] | None: ...
+
+class OrganizationConfigRepNotFound(OrganizationConfigRep): ...
+
+class OrganizationConfigRepUnknownStatus(OrganizationConfigRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
