+++ parsec/_parsec_pyi/protocol/ping.pyi
@@ -1,0 +1,48 @@
+# Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
+
+# Ping
+class AuthenticatedPingReq:
+    def __init__(self, ping: str) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def ping(self) -> str: ...
+
+class AuthenticatedPingRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> AuthenticatedPingRep: ...
+
+class AuthenticatedPingRepOk(AuthenticatedPingRep):
+    def __init__(self, pong: str) -> None: ...
+    @property
+    def pong(self) -> str: ...
+
+class AuthenticatedPingRepUnknownStatus(AuthenticatedPingRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class InvitedPingReq:
+    def __init__(self, ping: str) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def ping(self) -> str: ...
+
+class InvitedPingRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> InvitedPingRep: ...
+
+class InvitedPingRepOk(InvitedPingRep):
+    def __init__(self, pong: str) -> None: ...
+    @property
+    def pong(self) -> str: ...
+
+class InvitedPingRepUnknownStatus(InvitedPingRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
