+++ parsec/_parsec_pyi/protocol/realm.pyi
@@ -1,0 +1,369 @@
+# Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
+
+from typing import Iterable
+from parsec._parsec import DateTime, RealmID, DeviceID, UserID
+
+# Realm
+class RealmCreateReq:
+    def __init__(self, role_certificate: bytes) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def role_certificate(self) -> bytes: ...
+
+class RealmCreateRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> RealmCreateRep: ...
+
+class RealmCreateRepOk(RealmCreateRep): ...
+
+class RealmCreateRepInvalidCertification(RealmCreateRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmCreateRepInvalidData(RealmCreateRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmCreateRepNotFound(RealmCreateRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmCreateRepAlreadyExists(RealmCreateRep): ...
+
+class RealmCreateRepBadTimestamp(RealmCreateRep):
+    def __init__(
+        self,
+        reason: str | None,
+        ballpark_client_early_offset: float | None,
+        ballpark_client_late_offset: float | None,
+        backend_timestamp: DateTime | None,
+        client_timestamp: DateTime | None,
+    ) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+    @property
+    def ballpark_client_early_offset(self) -> float | None: ...
+    @property
+    def ballpark_client_late_offset(self) -> float | None: ...
+    @property
+    def backend_timestamp(self) -> DateTime | None: ...
+    @property
+    def client_timestamp(self) -> DateTime | None: ...
+
+class RealmCreateRepUnknownStatus(RealmCreateRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStatusReq:
+    def __init__(self, realm_id: RealmID) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def realm_id(self) -> RealmID: ...
+
+class RealmStatusRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> RealmStatusRep: ...
+
+class RealmStatusRepOk(RealmStatusRep):
+    def __init__(
+        self,
+        in_maintenance: bool,
+        maintenance_type: MaintenanceType | None,
+        maintenance_started_on: DateTime | None,
+        maintenance_started_by: DeviceID | None,
+        encryption_revision: int,
+    ) -> None: ...
+    @property
+    def in_maintenance(self) -> bool: ...
+    @property
+    def maintenance_type(self) -> MaintenanceType | None: ...
+    @property
+    def maintenance_started_on(self) -> DateTime | None: ...
+    @property
+    def maintenance_started_by(self) -> DeviceID | None: ...
+    @property
+    def encryption_revision(self) -> int: ...
+
+class RealmStatusRepNotAllowed(RealmStatusRep): ...
+
+class RealmStatusRepNotFound(RealmStatusRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStatusRepUnknownStatus(RealmStatusRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStatsReq:
+    def __init__(self, realm_id: RealmID) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def realm_id(self) -> RealmID: ...
+
+class RealmStatsRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> RealmStatsRep: ...
+
+class RealmStatsRepOk(RealmStatsRep):
+    def __init__(self, blocks_size: int, vlobs_size: int) -> None: ...
+    @property
+    def blocks_size(self) -> int: ...
+    @property
+    def vlobs_size(self) -> int: ...
+
+class RealmStatsRepNotAllowed(RealmStatsRep): ...
+
+class RealmStatsRepNotFound(RealmStatsRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStatsRepUnknownStatus(RealmStatsRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmGetRoleCertificatesReq:
+    def __init__(self, realm_id: RealmID) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def realm_id(self) -> RealmID: ...
+
+class RealmGetRoleCertificatesRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> RealmGetRoleCertificatesRep: ...
+
+class RealmGetRoleCertificatesRepOk(RealmGetRoleCertificatesRep):
+    def __init__(self, certificates: Iterable[bytes]) -> None: ...
+    @property
+    def certificates(self) -> tuple[bytes]: ...
+
+class RealmGetRoleCertificatesRepNotAllowed(RealmGetRoleCertificatesRep): ...
+
+class RealmGetRoleCertificatesRepNotFound(RealmGetRoleCertificatesRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmGetRoleCertificatesRepUnknownStatus(RealmGetRoleCertificatesRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmUpdateRolesReq:
+    def __init__(self, role_certificate: bytes, recipient_message: bytes | None) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def role_certificate(self) -> bytes: ...
+    @property
+    def recipient_message(self) -> bytes | None: ...
+
+class RealmUpdateRolesRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> RealmUpdateRolesRep: ...
+
+class RealmUpdateRolesRepOk(RealmUpdateRolesRep): ...
+
+class RealmUpdateRolesRepNotAllowed(RealmUpdateRolesRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmUpdateRolesRepInvalidCertification(RealmUpdateRolesRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmUpdateRolesRepInvalidData(RealmUpdateRolesRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmUpdateRolesRepAlreadyGranted(RealmUpdateRolesRep): ...
+
+class RealmUpdateRolesRepIncompatibleProfile(RealmUpdateRolesRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmUpdateRolesRepNotFound(RealmUpdateRolesRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmUpdateRolesRepInMaintenance(RealmUpdateRolesRep): ...
+class RealmUpdateRolesRepUserRevoked(RealmUpdateRolesRep): ...
+
+class RealmUpdateRolesRepRequireGreaterTimestamp(RealmUpdateRolesRep):
+    def __init__(self, strictly_greater_than: DateTime) -> None: ...
+    @property
+    def strictly_greater_than(self) -> DateTime: ...
+
+class RealmUpdateRolesRepBadTimestamp(RealmUpdateRolesRep):
+    def __init__(
+        self,
+        reason: str | None,
+        ballpark_client_early_offset: float,
+        ballpark_client_late_offset: float,
+        backend_timestamp: DateTime,
+        client_timestamp: DateTime,
+    ) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+    @property
+    def ballpark_client_early_offset(self) -> float: ...
+    @property
+    def ballpark_client_late_offset(self) -> float: ...
+    @property
+    def backend_timestamp(self) -> DateTime: ...
+    @property
+    def client_timestamp(self) -> DateTime: ...
+
+class RealmUpdateRolesRepUnknownStatus(RealmUpdateRolesRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStartReencryptionMaintenanceReq:
+    def __init__(
+        self,
+        realm_id: RealmID,
+        encryption_revision: int,
+        timestamp: DateTime,
+        per_participant_message: dict[UserID, bytes],
+    ) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def realm_id(self) -> RealmID: ...
+    @property
+    def encryption_revision(self) -> int: ...
+    @property
+    def timestamp(self) -> DateTime: ...
+    @property
+    def per_participant_message(self) -> dict[UserID, bytes]: ...
+
+class RealmStartReencryptionMaintenanceRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> RealmStartReencryptionMaintenanceRep: ...
+
+class RealmStartReencryptionMaintenanceRepOk(RealmStartReencryptionMaintenanceRep): ...
+class RealmStartReencryptionMaintenanceRepNotAllowed(RealmStartReencryptionMaintenanceRep): ...
+
+class RealmStartReencryptionMaintenanceRepNotFound(RealmStartReencryptionMaintenanceRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStartReencryptionMaintenanceRepBadEncryptionRevision(
+    RealmStartReencryptionMaintenanceRep
+): ...
+
+class RealmStartReencryptionMaintenanceRepParticipantMismatch(RealmStartReencryptionMaintenanceRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStartReencryptionMaintenanceRepMaintenanceError(RealmStartReencryptionMaintenanceRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmStartReencryptionMaintenanceRepInMaintenance(RealmStartReencryptionMaintenanceRep): ...
+
+class RealmStartReencryptionMaintenanceRepBadTimestamp(RealmStartReencryptionMaintenanceRep):
+    def __init__(
+        self,
+        reason: str | None,
+        ballpark_client_early_offset: float,
+        ballpark_client_late_offset: float,
+        backend_timestamp: DateTime,
+        client_timestamp: DateTime,
+    ) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+    @property
+    def ballpark_client_early_offset(self) -> float: ...
+    @property
+    def ballpark_client_late_offset(self) -> float: ...
+    @property
+    def backend_timestamp(self) -> DateTime: ...
+    @property
+    def client_timestamp(self) -> DateTime: ...
+
+class RealmStartReencryptionMaintenanceRepUnknownStatus(RealmStartReencryptionMaintenanceRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmFinishReencryptionMaintenanceReq:
+    def __init__(self, realm_id: RealmID, encryption_revision: int) -> None: ...
+    def dump(self) -> bytes: ...
+    @property
+    def realm_id(self) -> RealmID: ...
+    @property
+    def encryption_revision(self) -> int: ...
+
+class RealmFinishReencryptionMaintenanceRep:
+    def dump(self) -> bytes: ...
+    @classmethod
+    def load(cls, buf: bytes) -> RealmFinishReencryptionMaintenanceRep: ...
+
+class RealmFinishReencryptionMaintenanceRepOk(RealmFinishReencryptionMaintenanceRep): ...
+class RealmFinishReencryptionMaintenanceRepNotAllowed(RealmFinishReencryptionMaintenanceRep): ...
+
+class RealmFinishReencryptionMaintenanceRepNotFound(RealmFinishReencryptionMaintenanceRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmFinishReencryptionMaintenanceRepBadEncryptionRevision(
+    RealmFinishReencryptionMaintenanceRep
+): ...
+
+class RealmFinishReencryptionMaintenanceRepNotInMaintenance(RealmFinishReencryptionMaintenanceRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmFinishReencryptionMaintenanceRepMaintenanceError(RealmFinishReencryptionMaintenanceRep):
+    def __init__(self, reason: str | None) -> None: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class RealmFinishReencryptionMaintenanceRepUnknownStatus(RealmFinishReencryptionMaintenanceRep):
+    def __init__(self, status: str, reason: str | None) -> None: ...
+    @property
+    def status(self) -> str: ...
+    @property
+    def reason(self) -> str | None: ...
+
+class MaintenanceType:
+    def __init__(self, maintenance_type: str) -> None: ...
+    @classmethod
+    def GARBAGE_COLLECTION(cls) -> MaintenanceType: ...
+    @classmethod
+    def REENCRYPTION(cls) -> MaintenanceType: ...
